644678b5ddb4a Support Subform: Transfer Item further work The transfer device is mostly working on the production site, as you might have already realized, but... single zone dont have an account, disabled checkbox SF154 - if account - transfer it no account - dont have to worry It changes the ownership of the device, but not the user account...even if transfer account is checked Go by this form: /_spectros/transfer_ownership?expandNav=9
SF4154 - needs an account to talk to the monitor portal
look at transfer device form
monitors have automation account not human
transfer the automation account - devices login to the portal to upload data

632c771fea875 Support Action Ship Unit Subform
close the action after ship
TCKT150 - should be:  TCKT000150 - find the ticket number function
return sprintf("%06d",$this->id);
warehouse shipping page, make sure that the item is added to the package

shipping_packages - BUG: package_id is NULL
seems like the new shipping action above isnt recording the monitor in the shipment
Fix page format - /_shipping/admin_shipment?id=24

632c771fea875 Support Action Ship Unit Subform
In the /_support/action page, *if* the action type is Ship Unit,
display a list of existing shipments for the customer that are in NEW status.
If the user clicks one of these, this item will be added to the shipment.
we need to SHIP  not recieve - create a new shipment
or add to existing shipment, but only if Spectros is already ongoing
Existing SPECTROS Shipments

Create a new Shipment
add a pull down for shipping vendor and a tracking number field to create a new shipment.
The shipment must be associated with the ticket, and have an address from the RMA, or the Sales Order if a fulfillment ticket.
have a check for item in shipment to not show the buttons again?
Add item to Shipment

SF12TEST01
Create an event shipment, then close the task
shipping_shipments - document_number = TCKT[TICKET NUMBER]
add to shipment bug